requirements.txt
attrs==19.3.0
autopep8==1.5.3
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
codecov==2.1.8
coloredlogs==14.0
coverage==5.2.1
cryptography==3.0
flake8==3.8.3
flake8-print==3.1.4
flake8-quotes==3.2.0
humanfriendly==8.2
idna==2.10
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pip-autoremove==0.9.1
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
PyYAML==5.3.1
requests==2.24.0
rope==0.17.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
wcwidth==0.2.5
